WIP split off session module
- replace htmlement with html5lib; more predictable results
- use platform-specific directories for config and state
- define and use program name and version
