+++ src/tyora/__main__.py
@@ -1,0 +1,6 @@
+from sys import exit
+
+from . import tyora
+
+if __name__ == "__main__":
+    exit(tyora.main())
